삭제 될 브랜치

+++ src/main/java/sejongPromise/backend/domain/review/model/Review.java
@@ -34,7 +34,7 @@ public class Review extends BaseEntity {
     @ManyToOne(fetch = FetchType.LAZY)
     private Student student;
 
-    private Integer score;
+    private Float score;
 
     private Integer volume;
 
@@ -45,7 +45,6 @@ public class Review extends BaseEntity {
     private ReviewStatus status;
     @Enumerated(EnumType.STRING)
     private Semester passSemester;
-
     @OneToMany(mappedBy = "review", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<Report> reports = new ArrayList<>();
 
@@ -53,7 +52,7 @@ public class Review extends BaseEntity {
     @Builder
     private Review(@NonNull Book book,
                    @NonNull Student student,
-                   @NonNull Integer score,
+                   @NonNull Float score,
                    @NonNull Integer volume,
                    @NonNull BookRatio ratio,
                    String comment) {
@@ -67,7 +66,7 @@ public class Review extends BaseEntity {
         this.passSemester = student.getSemester();
     }
 
-    public void updateReview(Integer score, Integer volume, BookRatio ratio, String comment) {
+    public void updateReview(Float score, Integer volume, BookRatio ratio, String comment) {
         this.score = score;
         this.volume = volume;
         this.ratio = ratio;
